third_party/WebKit/LayoutTests/fast/table/border-recalc.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<table>
<tbody>
<tr><td>
    <div></div>
    <div></div>
    <div></div>
    <div></div>
    <div></div>
    <div></div>
    <div></div>
    <div></div>
    <div></div>
    <div></div>
</td></tr>
</tbody>
</table>
<script>
description("Check that we don't recalc table cell contents when presentation attributes change.");

document.body.offsetTop; // Force layout.
document.getElementsByTagName("table")[0].setAttribute("border", "2");
if (window.internals)
    shouldBe("internals.updateStyleAndReturnAffectedElementCount()", "4");
</script>
